Use utf-8 instead of us-ascii. Catch bad program runs.
At this point, if actual utf-8 is in the message, we are violating rfc 3156, since the message is no longer 7-bit.
(see section 3). We either need to make this BASE64 (very wasteful) or try to go quoted printable.

Note also, that according to section 5, trailing whitespace is to be removed from every line, but
that appears to be *after* the transfer encoding (e.g. quoted printable), so we *ought* to be okay.