Experiments with changing prompt-toolkit's eventloop.
So far, so good. We're able to have libuv handle events and forward them to prompt-toolkit.
Ideally, we need a way to have interruptions be processed differently (e.g. actually get
signals)
Ideally, the prompting and result would be asynchronous (e.g. show prompt, run callback
when prompt is complete).
This might fall apart a bit if a new command prompt shouldn't be drawn (e.g. we spawn
a child to be a viewer, like less, or an editor, like vim, but we still want our event loop
running to keep the IMAP connection open, but at that point don't want to write to the terminal).