Experiment to see if libuv remembers handlers between a stop and start.
Turns out it does. As such, we probably don't need to do crazy things with the
prompt. When we ask for a prompt, our machine is running. Within an internal
command launched as the result of a prompt (e.g. fetching a message), we can
prime the IMAP system with callbacks that watch for the data we want, then run
the libuv loop. The callback can stop the loop when it is done, allowing our
outer loop to show another prompt, which in turn resumes the event loop.
Magic!