Move code processing to common function. Make debug statements flagged.
This cleans up some of the processing, especially in the connection phase, and
also now captures capability changes resulting from login that the server
provices for us.

The debug info can still be useful, but isn't often needed and clutters the
screen. Making it depend on a debug variable cleans it up for now. Eventually,
it should be calling some kind of logging function with a debug level so
filtering can be done by the logger instead of us.

Also added is processing for untagged data without an index number prefix
(e.g. untagged FLAGS response).

At this point, connecting to local dovecot is clean.