Correct some issues with the parser.
First, allow a literal's size tage to begin without a preceding space (e.g.
immediately after an open parenthesis).

Second, read the actual RFC syntax for a literal instead of guessing (we had
which CRLF wasn't part of the count wrong. Sometimes literals don't end in
CRLF, they just usually do, and when they do, it is actually part of the
literal (if it is within the count after the initial CRLF).