Add show and structure to using new imap lib. Also encoding fixes.
Encoding is temporary and wrong. Outlook says it is using iso-8895-1, but actually is using windows-1252. Trying to interpret
as iso-8895-1 either leads to errors or control characters in output.

For now, we are hard coded to assume windows-1252, but this would of course break if someone sent us anything else that isn't a subset.
For example, a utf-8 message that wasn't actually ASCII would break.

Also, I'm assumint utf-8 output, which isn't true for some terminals.