Convert header printing to use format instead of % for formatting.
This is the Python 3 way. Out with the old, in with the new.

Part of doing this is to provide more flexible formatting for the header setting. With format, we can
get parts of objects and object specific rules (such as date formatting), without losing anything
from the old method.

This will allow custom date string display without having to have two settings, and should make color
and other such things easier.