add 'z' and 'Z' commands.
Note: Z is a bit buggy. In particular, by adding the current page to the list, we
can no longer tell if the current page exceeds the bounds of the pages we care about.
Thus, if the current page is below the first interesting one and 'Z -' is given, or it
is after the last interesting page and 'Z' or 'Z +' is given, we reshow the uninteresting
page.