Make most message list activity common. Don't exit on most exceptions.
This introduces several decorators for commands to show exception information instead of aborting and
for handling the repeditive argument conversions.

This has introduced new repeditive code, so that still needs to be worked out.

The behavior of the commands as far as selection goes should be much more consistent now, and closer
to what mailx does.