Partial support for virtual folders.
This allows one to specify a message list and convert the view of messages
to consist of just that list.

This allows one to, for example, set the list of flagged messages as the virtual
folder and then move about that list using normal mailx navigation (e.g. just
pressing enter shows the next message in the list/virtual folder rather than
the next message in the real folder, which might not match your criteria).

Currently only the h, z, Z, and print commands are supported (and by
virtue of print, just pressing enter for next message works too).

f, show, flag and unflag don't yet. I haven't decided if vf itself should or not.

Additionally, the message list interpreter doesn't support it yet, which is very
broken. Using plain numbers and ranges is fine, but selectors like :u and :f will
use the real folder numbers even though some commands aren't. This is okay for f, but
disasterous for print or used on its own.

Still, useful enough for me right now to make it commit-able.