First step to validating SSL connections.
Unfortunately, Ubuntu 14.04 LTS ships a python that predates the SSLContext stuff that makes
some of this easier, so we must manage ourselves for now.

This needs cleanup; we shouldn't be setting values directly on imap4, and it shouldn't expect it.
We should also allow ca override per hostname, and SSL validity check override per hostname.

For example, maybe a user doesn't want to validate for localhost connections, or they
trust a particular self-signed cert, but only for a particular host.