Fixup some cases with unicode handling.
So far, this version hasn't died on me like the others. Pretty sure we aren't handling this
properly still, though.

The Python email classes seem to try to do a lot to keep encodings handled correctly, but
I haven't found a clean way to deal with it for editing the message and headers. Likewise,
getting the data out without having the lib do the printing on our behalf isn't obvious
to me yet.