Experiment with gpgme
Includes two incompatible demos. One generates a signature only, the other checks
a message with signature in one stream (that is, not what the first one generates).

So far this looks do-able for implementing an OpenPGP/MIME maker/checker, better
than calling gpg directly, since this way I can find out what hash was used, which
is apparently needed in one of the headers.