Enable host-name checking on systems that support it.
Also, move the warning about lack of checking to only display when the host doesn't support it.

So far, we are able to do host checking and easily disable old algorithms on Ubuntu 16.04. We are not
able to on Ubuntu 14.04. It looks like we should be able to in the current Centos 7 (but untested).