Fix bug where not all data made it to external viewer.
The write fills the OS pipe and then procedes in code (non-blocking). PyUV will continue writing the
buffer for us as the pipe becomes writable. However, after it returned the first time, we marked it
for closing by calling close, which results in it *not* writing more to the buffer as the pipe becomes
writable. Oops.

Setting up a callback to close it on completion or error makes everything happy. We'll ignore the
error (if any) on the pipe for now; we only really care about the program we executed.