Add MessageList type to simplify some of the processing.
In several places (mostly when trying to show headlines) we were passing around strings representing
what we wanted to really show, roughly in the format that IMAP wanted. Then we tried to do
some processing on that (e.g. for virtual folders). The result got... messy.

This creates an interface for managing message lists and provides some room for improvement without
requiring that we rewrite a lot of the code to do it.