Set a title.
In the future, we might have a setting for it so that it can include information
like number of unread messages or whatever.
For now, this mostly saves my title from the repeated expansion that
occurs everytime I enter/exit Vim for editing a message, where Vim tries
to restore the old title, but doesn't understand that I'm running tmux, so
every time it does, I get another copy of the tmux header prepended.