Fix bug in processing mailcap commands with quotes.
We were just splitting on whitespace, so quotes were passed verbatim to the
underlying command, which then would try, for example, to open a file that
included quotes in the filename.

We don't have to worry about actual quotes in the filename since the files
are generated with temporary names, rather than user data.