Fix header mangling when reply to a message with Reply-to header.
Most messages just use the From field by itself, and we can reply to those just fine.
However, some messages also specify a Reply-to header, and we were mangling that when
trying to insert it into the To list. Wrapped it in an array like we do for the From field so the insertion
works properly.