Allow user to give a shell command for auth instead of using keyring.
Inspiration for this came from the luks configuration of /etc/crypttab
which has a 'keyscript' setting, which I use to have gpg decrypt a long
random key. I thought it would be good to have something similar for
my mail program.

Just before implementing it, I noticed that s-nail has taken up roughly
where heirloom-mailx left off. Ubuntu switched from heirloom-mailx to s-nail
somewhere between 14.04 and 16.04. I also saw in s-nail's man page that
they already implement this great feature, but they call it
'agent-shell-lookup' instead of 'keyscript'. I was going to follow crypttab,
but since some 'mailx' users already expect agent-shell-lookup, I'll use that
naming for this.

As of this commit, it works for simple scripts, and it is slightly more
specify-able than s-nail (this allows different scripts based on protocol and
port, which might be useful if pop vs imap use different auth, or imap vs
imaps (not so likely), or if you connect to a machine that forwards different
ports to different machines (more likely. For example, a reverse proxy ssl
termination point that passes connections to different machines in the backend).

What doesn't work yet, is I don't setup any environment variables, so existing
s-nail scripts that rely on NAIL_USER and NAIL_HOST won't work. I'm not sure
yet if our vars should be 'NAIL*' for compat, or 'MAILNEX*' to not collide.