Decode headers during print.
This improves the case where, for example, a subject or full name has utf-8 encoded
characters in it. Previously we decoded the From header for the header lines (list
of emails), but then didn't in the print view.

Print (capital P) still doesn't decode. 'show' also doesn't, but also shouldn't.