First pass PGP signature checking.
Very inefficient. In particular, parts of the message are retrieved twice:
once for signature checking, and again for display.

Information display about the signature is limited, and also only shown in
the structure view when 'print' or 'Print'ing the message (that is, it isn't
checked with the structure command, nor with 'show'). If 'noshowstructure' setting,
then the signature is checked, but the result isn't shown anywhere.

Only the primary uid of the signer is shown. It is not checked against the sender, and
no header matching is performed.

If there are multiple signers, only the last verified good or bad signature (by gpgme) is shown.