Gracefully exit a given mode on ctrl+c (or equiv).
Previously, an interrupt would exit all of mailnex, no matter when. Now, it
will exit the message editor or attachment editor instead of everything at
once, which is more like mailx. Our current worst case is in the attachment
editor, ctrl+c returns to the message editor, where ctrl+c returns to regular
mailnex, and one more ctrl+c exits.

It should be noted that we only catch this while waiting for input. It is
conceivable that an interrupt could occur elsewhere, e.g. just after a user
presses 'enter' and before we start getting the next line, which would trigger
the outer "exit mailnex" handler. I *think* prompt_toolkit leaves the terminal
in raw mode, and generates the KeyboardInterrupt itself, so this is less
likely (signal would have to be from kill() to get past that). Either way,
this is probably what a user expects; if we are incredibly slow, hitting
ctrl+c a lot will probably cause us to exit.