Attempt to avoid exception loops when cli breaks.
Occasionally we can get into a situation where the cli is in a bad state. If we continue
feeding it characters, it raises an exception (e.g. StopIteration) for every character. Trying
to exit with ctrl+d, ctrl+c, or even ctrl+\ doesn't work, because the terminal is in raw and
the cli isn't processing those characters.

If we close off the stream on an exception, things seem better. Things can be re-opened later.