Fix issue where signatures were invalid with long headers.
The python email library tried to help by not wrapping headers if it detected a signed message.
However, it changing based on wrapping actually made things worse for us, so we just override that.

When we check signatures, we are checking against the (slightly reformatted per RFC) raw message,
so the email package won't affect us there anyway.