First pass displaying encrypted messages.
Not a great implementation, but so far it is working. Utilizes a cache to allow decrypting
once during structure parsing. Unfortunately, it doesn't automatically clear the cache,
so decrypted data stays in RAM until the program exits (and possibly after), and yet
we don't consult the cache next time we parse, so the data will get decrypted again
if the user asks for the message again.