Clean empty recipient addresses.
By default, our tab completion ends with a comma. This causes the sending logic to
include an empty recipient. Some systems promote this to 'root' or 'nobody' which
can either deliver a message to an unintended recipient, or give the sender a
confusing delivery failure message.

This could also possibly be fixed by preventing trailing commas in the first place,
but still someone could type a leading, trailing, or double comma directly, and
the outcome is not likely what was intended.