Use separate database for each box.
This was the best I could come up with for now. Everything else had
difficulties when one wants search results limited to a particular box
(folder). For example, adding the location as a field (like we do with
subject) was unreliable for filtering. Doing manual filtering after xapian
would require storing data (which is included with this patch) and somehow
mapping our result numbers to the xapian result numbers for the search
continuance (and repeated looping to get enough results when some are filtered
out).

IIRC, xapian has a way to join DBs for a query of multiple sources, so we'll
probably try going that way at some point for cross-box search.