WIP: threading view.
Currently spews a lot of debug info, is slow, especially over a non-local link,
and doesn't update as new messages arrive. Abuses virtual folder mode. Leaves out
a lot of messges. Only partially follows references headers. In short, leaves much
to be desired, but it is a start that mostly works for reading messages in a thread
in order.