WIP: better threading. Also RFC 5256 summary.
We now handle messages who are replies but we didn't find the parent. Also, we now can do
top-sorting by most-rescent thread-member (by sequence number, not by date yet).

This also includes a summary form, slightly reinterpreted from the RFC on sorting and threading.
In particular, the REFERENCES threading algorithm has different step numbers because some
of the steps in the RFC weren't numbered. The RFC algorithm is the goal to reach, but
this is thus far working well enough for my use.