Add sanatizer for headlines.
I was recently shocked to see emails with new lines in the subject. They messed up
my nice headline display by making the message occupy multiple lines. I also
realized that, since the headlines (h, z, and Z commands) aren't handled by
a pager, I couldn't rely on the pager to clean the output for me (like 'less'
does), so a malicious email could contain control characters to change the appearance
of my terminal, or (depending on the terminal) mount an attack by overwriting a file.

This first pass just cleans up the headlines. Later, we should also do this to all
displayed text so that we don't rely on the pager for safety.