Support arbitrary Address Families (IPv6 support).
Previously, IPv4 sockets were all we could handle. Now, we use the generic interface
to go from name resolution to connection. We should thus be able to connect
to any stream socket over anything. In practice, this just means that both IPv4 and IPv6 work.

Also, unlike before, we now try all addresses given by name resolution instead of just the
first, so if a site has multiple A records and the first host is down, we can still
connect via the second (or third or whatever).

During this, we will only attempt to authenticate with the first server we can connect to.
This means that, for example, a server which is up and accepting connections but cannot
communicate to its backend authentication server will cause us to fail to login, even
if a second server could have worked.

If we were instead to continue to the next server after auth with the first
server failed, we could end up in a situation where a mistyped password
results in account lock-out as we try the bad credentials against multiple
servers, adding to the failed auth count.

Since account lock-out is probably worse than failing, I chose the former
implementation over the latter.