Return to message editing on error sending.
Previously, an error from composing would dump the message to ~/dead.letter
and leave you back at the main prompt (or worse, for some errors it would drop
the message completely leaving one with no recovery).

That we also don't support the '~d' editor command to read the dead.letter
file makes this particularly frustrating. A user would need to launch an
editor (e.g. '~v') and use that to recover most of the letter. That doesn't
help with file attachment lists nor pulling in the recipient list.

This change makes things much better, especially for SMTP users, where a typo
in a recipient address was very bad.