Fix error when attaching binary data.
There was an error in the code that checked to see if data was outside of
normal printable ASCII to decide if an attachment should be base64 encoded.
Because we now have string literals as unicode, the whitelisted whitespace was
treated as unicode, making the check try to decode ASCII into a unicode string
for comparison to the unicode whitespace, which of course breaks if you have
data outside of ASCII (as you likely do with binary files).

Force the whitespace list to be a bytes-string, and all is well.