imap4: fix "nomatch" on idle complete warnings
Make processing until a tagstr a function. Then, doSimpleCommand can issue a
'done', process until the idle is done, then issue the command it wants to do,
and wait for that tag to complete.

I could alternatively have called doSimpleCommand recursively, but having the
parsing of data be separate from the issuance of commands makes it a bit
easier to read, and may make it easier to convert this to asynchronous code
later.