Improve handling of split names when replying.
Previously, when other mail clients split a To or Cc email address
across lines within the name part and the name part was quoted, we would
get a very mangled name in our To and/or Cc header.

This change results in the address itself staying valid but probably
getting extra whitespace in the name during the replication. Should be
improved, but at least this is less broken.